global mock for email/contract
