+++ apps/admin/jest.config.js
@@ -2,5 +2,9 @@ const defaults = require('../../jest.config');
 
 module.exports = {
   ...defaults,
-  moduleDirectories: ['node_modules', '../../node_modules', 'app']
+  moduleDirectories: ['node_modules', '../../node_modules', 'app'],
+  roots: [
+    '<rootDir>/app',
+    defaults.roots[1],
+  ],
 };
