--- apps/admin/manager-ts.code-workspace
@@ -1,23 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		},
-		{
-			"path": "../../shared-components/src"
-		}
-	],
-  "settings": {
-    "files.associations": {
-      ".babelrc": "jsonc",
-      ".eslintrc": "jsonc",
-      ".prettierrc": "jsonc",
-      ".stylelintrc": "json",
-      ".dockerignore": "ignore",
-      ".eslintignore": "ignore",
-      ".flowconfig": "ignore"
-    },
-    "typescript.tsdk": "..\\node_modules\\typescript\\lib",
-    "deepcode.review.results.hideInformationIssues": false
-  }
-}
