+++ libs/contract/package.json
@@ -29,7 +29,6 @@
     "clean:contracts": "rimraf ./src/contracts",
     "clean:types": "rimraf ./src/types && rimraf ./migrations/types",
     "test:contract": "truffle test",
-    "test": "yarn jest",
     "dev:live": "cross-env CONFIG_NAME=devlive yarn dev:live:migrate",
     "dev:live:emulator": "start yarn truffle develop",
     "dev:live:migrate": "yarn wait-port 9545 && yarn truffle migrate --network devlive",
