--- libs/tx-gmail/src/gmail.json
@@ -1,1 +1,0 @@
-{"installed":{"client_id":"34289122317-ts71pej3nu2ejpedf0rbsvigedndq348.apps.googleusercontent.com","project_id":"quickstart-1559424267516","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"CXTB-V_TndVfcljqcJhiJo_J","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
