--- libs/tx-processing/src/__mocks__/@thecointech/email.ts
@@ -1,4 +1,0 @@
-// Uncomment to see outputs if things are breaking
-export function SendMail(_subject: string, _message: string) {
-  //console.log(`Sending Mail: \nsubject: ${subject}\nmessage:\n${message}`);
-}
